+++ internal/stock/domain.go
@@ -11,8 +11,8 @@ import (
 type Request struct {
 	InventoryID      string                       `validate:"required" json:"inventoryID"`
 	SupplierID       string                       `validate:"required" json:"supplierID"`
-	PurchasePrice    float32                      `validate:"required" json:"purchasePrice"`
-	PurchaseQuantity float32                      `validate:"required" json:"purchaseQuantity"`
+	PurchasePrice    float64                      `validate:"required" json:"purchasePrice"`
+	PurchaseQuantity float64                      `validate:"required" json:"purchaseQuantity"`
 	PurchaseUnit     usageUnitModule.UsageUnitReq `validate:"required" json:"purchaseUnit"`
 	Remark           string                       `json:"remark"`
 }
@@ -23,8 +23,8 @@ type Prototype struct {
 	Inventory        *inventoryModule.Prototype `json:"inventory"`
 	SupplierID       string                     `json:"supplierID"`
 	Supplier         *supplierModule.Prototype  `json:"supplier"`
-	PurchasePrice    float32                    `json:"purchasePrice"`
-	PurchaseQuantity float32                    `json:"purchaseQuantity"`
+	PurchasePrice    float64                    `json:"purchasePrice"`
+	PurchaseQuantity float64                    `json:"purchaseQuantity"`
 	PurchaseUnit     usageUnitModule.Prototype  `json:"purchaseUnit"`
 	Remark           string                     `json:"remark"`
 	CreatedAt        *time.Time                 `json:"createdAt"`
